perf: cache callbacks
